+++ shellkit/kit_test.go
@@ -1,7 +1,0 @@
-package shellkit
-
-import "testing"
-
-func TestExecute(t *testing.T) {
-	Execute("cmd", "/c", "dir")
-}