+++ changes.md
@@ -1,3 +1,7 @@
 # Changes
 - use go mod
 - add ProxyClient method
+
+# Changes on 30 Aug 2024
+- remove shellkit. Use magefile's sh package instead
+- replace ullive IOProgress bar with progressbar, which seems like cause hung in go 1.23